Fix type cast for elts

+++ benches/fibonacci_codec.rs
@@ -52,7 +52,7 @@ fn encode_multiple_benchmark(c: &mut Criterion) {
         },
         ALL,
     )
-    .throughput(|_s| Throughput::Elements(ELTS as u32));
+    .throughput(|_s| Throughput::Elements(ELTS as u64));
     c.bench(id, bm);
 }
 
@@ -67,7 +67,7 @@ fn decode_multiple_benchmark(c: &mut Criterion) {
         },
         ALL,
     )
-    .throughput(|_s| Throughput::Elements(ELTS as u32));
+    .throughput(|_s| Throughput::Elements(ELTS as u64));
     c.bench(id, bm);
 }
